refactor(store): extract chord overlay and scale visibility into dedicated modules
Split the monolithic atoms.ts (~1700 lines) into four focused domain modules:

- fingeringAtoms.ts: FingeringPattern, cagedShapesAtom, fingeringPatternAtom,
  toggleCagedShapeAtom, selectSingleCagedShapeAtom, npsPositionAtom,
  clickedShapeAtom, recenterKeyAtom
- scaleAtoms.ts: rootNoteAtom, scaleNameAtom, scaleBrowseModeAtom,
  accidentalModeAtom, useFlatsAtom, scaleNotesAtom, colorNotesAtom,
  activeBrowseOptionAtom, scaleLabelAtom, degreeChipsAtom, hiddenNotesAtom
- chordOverlayAtoms.ts: chord base atoms, practiceLensAtom, focusPresetAtom,
  viewModeAtom, customMembersAtom, hideNonChordNotesAtom, all chord derived
  atoms including allChordMembersAtom
- practiceLensAtoms.ts: noteSemanticMapAtom, practiceCuesAtom,
  showChordPracticeBarAtom, practiceBarColorNotesAtom,
  practiceBarTitleAtom/BadgeAtom, practiceBarShared/OutsideMembersAtom

Also promoted shared storage adapters (rawStringStorage, booleanStorage,
constrainedNumberStorage, GET_ON_INIT) to utils/storage.ts and moved
migrateLegacyKeys() there so it fires before any atomWithStorage({ getOnInit:
true }) reads — fixing a module-load-order race the old single-file approach
avoided naturally.

atoms.ts is now a thin re-export barrel + remaining atoms (display, tuning,
UI tabs, shape data, shape-local, summary, actions). All existing import paths
from "../store/atoms" continue to work unchanged.
