+++ src/store/practiceLensAtoms.ts
@@ -1,0 +1,334 @@
+import { atom } from "jotai";
+import {
+  NOTES,
+  ENHARMONICS,
+  INTERVAL_NAMES,
+  getScaleNotes,
+  getNoteDisplay,
+  formatAccidental,
+} from "../theory";
+import type {
+  ChordMemberName,
+  NoteSemantics,
+  PracticeCue,
+  PracticeCueNote,
+  ChordRowEntry,
+  PracticeBarColorNote,
+} from "../theory";
+import {
+  rootNoteAtom,
+  scaleNameAtom,
+  scaleNotesAtom,
+  colorNotesAtom,
+  useFlatsAtom,
+} from "./scaleAtoms";
+import {
+  chordRootAtom,
+  chordTypeAtom,
+  chordLabelAtom,
+  practiceLensAtom,
+  focusPresetAtom,
+  hasOutsideChordMembersAtom,
+  activeChordMembersAtom,
+  activeChordTonesAtom,
+  allChordMembersAtom,
+} from "./chordOverlayAtoms";
+
+// Guide tone member names: 3rd and 7th (before formatAccidental)
+const GUIDE_TONE_RAW = new Set(["b3", "3", "b7", "7"]);
+// After formatAccidental: "b3"→"♭3", "b7"→"♭7"
+const GUIDE_TONE_FORMATTED = new Set(["♭3", "3", "♭7", "7"]);
+
+// ---------------------------------------------------------------------------
+// Practice bar color notes — derived from scale color tones
+// ---------------------------------------------------------------------------
+
+export const practiceBarColorNotesAtom = atom((get) => {
+  const colorNotes = get(colorNotesAtom);
+  const rootNote = get(rootNoteAtom);
+  const useFlats = get(useFlatsAtom);
+
+  if (colorNotes.length === 0) return [] as PracticeBarColorNote[];
+  const rootIdx = NOTES.indexOf(rootNote);
+  if (rootIdx === -1) return [] as PracticeBarColorNote[];
+  return colorNotes.map((note) => {
+    const noteIdx = NOTES.indexOf(note);
+    const interval = (noteIdx - rootIdx + 12) % 12;
+    const intervalName = INTERVAL_NAMES[interval] ?? "";
+    return {
+      internalNote: note,
+      displayNote: formatAccidental(getNoteDisplay(note, rootNote, useFlats)),
+      intervalName: formatAccidental(intervalName),
+    };
+  });
+});
+
+export const practiceBarColorNotesFilteredAtom = atom((get) => {
+  const activeChordTones = get(activeChordTonesAtom);
+  const practiceBarColorNotes = get(practiceBarColorNotesAtom);
+  const chordToneSet = new Set(activeChordTones);
+  return practiceBarColorNotes.filter((n) => !chordToneSet.has(n.internalNote));
+});
+
+// ---------------------------------------------------------------------------
+// Note semantic map — composable properties per note (multiple can coexist)
+// ---------------------------------------------------------------------------
+
+/**
+ * Returns a map of note → NoteSemantics where multiple boolean properties can
+ * coexist on one note. Crucially, a chord root that is outside the scale will
+ * have both isChordRoot=true and isTension=true — something the old single-role
+ * enum (NoteRole) could not represent.
+ */
+export const noteSemanticMapAtom = atom((get) => {
+  const chordType = get(chordTypeAtom);
+  if (!chordType) return new Map<string, NoteSemantics>();
+
+  const rootNote = get(rootNoteAtom);
+  const scaleName = get(scaleNameAtom);
+  const chordRoot = get(chordRootAtom);
+  const activeChordMembers = get(activeChordMembersAtom);
+  const colorNotes = get(colorNotesAtom);
+
+  const scaleNoteSet = new Set(getScaleNotes(rootNote, scaleName));
+  const colorNoteSet = new Set(colorNotes);
+
+  const memberByNote = new Map<string, (typeof activeChordMembers)[number]>();
+  for (const m of activeChordMembers) memberByNote.set(m.note, m);
+  const activeChordToneSet = new Set(activeChordMembers.map((m) => m.note));
+
+  const map = new Map<string, NoteSemantics>();
+  for (const note of NOTES) {
+    const isInScale = scaleNoteSet.has(note);
+    const isChordRoot = note === chordRoot;
+    const isChordTone = activeChordToneSet.has(note);
+    const enh = ENHARMONICS[note];
+    const isColorTone = colorNoteSet.has(note) || (!!enh && colorNoteSet.has(enh));
+    const member = memberByNote.get(note);
+    const isGuideTone = !!(member && GUIDE_TONE_RAW.has(member.name));
+    const isTension = isChordTone && !isInScale;
+    const isScaleRoot =
+      note === rootNote ||
+      ENHARMONICS[note] === rootNote ||
+      ENHARMONICS[rootNote] === note;
+
+    if (isInScale || isChordTone || isColorTone) {
+      map.set(note, {
+        isScaleRoot: !!isScaleRoot,
+        isChordRoot,
+        isChordTone,
+        isInScale,
+        isColorTone,
+        isGuideTone,
+        isTension,
+        memberName: member?.name as ChordMemberName | undefined,
+      });
+    }
+  }
+  return map;
+});
+
+// ---------------------------------------------------------------------------
+// Practice cues — coaching lines derived from the active practice lens
+// ---------------------------------------------------------------------------
+
+/**
+ * Derives ordered coaching cues for the practice bar based on the active lens.
+ * Each cue has a label ("Land on", "Guide tones", "Color note", "Tension") and
+ * a list of notes with styling hints and optional resolution targets.
+ */
+export const practiceCuesAtom = atom((get) => {
+  const chordType = get(chordTypeAtom);
+  if (!chordType) return [] as PracticeCue[];
+
+  const practiceLens = get(practiceLensAtom);
+  const chordRoot = get(chordRootAtom);
+  const useFlats = get(useFlatsAtom);
+  const allChordMembers = get(allChordMembersAtom);
+  const colorNotesFiltered = get(practiceBarColorNotesFilteredAtom);
+  const scaleNotes = get(scaleNotesAtom);
+
+  const displayNote = (note: string) =>
+    formatAccidental(getNoteDisplay(note, chordRoot, useFlats));
+
+  const toCueNote = (e: ChordRowEntry): PracticeCueNote => ({
+    internalNote: e.internalNote,
+    displayNote: e.displayNote,
+    intervalName: e.memberName,
+    role: e.role,
+  });
+
+  // Find the nearest in-scale note (half-step up or down) as a resolution target.
+  const findResolution = (
+    note: string,
+  ): { internalNote: string; displayNote: string } | undefined => {
+    const noteIdx = NOTES.indexOf(note);
+    if (noteIdx === -1) return undefined;
+    const scaleNoteSet = new Set(scaleNotes);
+    const up = NOTES[(noteIdx + 1) % 12];
+    const down = NOTES[(noteIdx + 11) % 12];
+    const resolved = scaleNoteSet.has(up)
+      ? up
+      : scaleNoteSet.has(down)
+        ? down
+        : undefined;
+    if (!resolved) return undefined;
+    return { internalNote: resolved, displayNote: displayNote(resolved) };
+  };
+
+  const cues: PracticeCue[] = [];
+
+  switch (practiceLens) {
+    case "targets": {
+      if (allChordMembers.length > 0) {
+        cues.push({
+          kind: "land-on",
+          label: "Land on",
+          notes: allChordMembers.map(toCueNote),
+        });
+      }
+      break;
+    }
+
+    case "guide-tones": {
+      const guideNotes = allChordMembers.filter((e) =>
+        GUIDE_TONE_FORMATTED.has(e.memberName),
+      );
+      if (guideNotes.length > 0) {
+        cues.push({
+          kind: "guide-tones",
+          label: "Guide tones",
+          notes: guideNotes.map((e) => ({
+            ...toCueNote(e),
+            role: "guide-tone" as const,
+          })),
+        });
+      } else {
+        // Fallback for power chords (no 3rd/7th): show all chord tones.
+        cues.push({
+          kind: "land-on",
+          label: "Land on",
+          notes: allChordMembers.map(toCueNote),
+        });
+      }
+      break;
+    }
+
+    case "color": {
+      if (colorNotesFiltered.length > 0) {
+        cues.push({
+          kind: "color-note",
+          label: colorNotesFiltered.length === 1 ? "Color note" : "Color notes",
+          notes: colorNotesFiltered.map((n) => ({
+            internalNote: n.internalNote,
+            displayNote: n.displayNote,
+            intervalName: n.intervalName,
+            role: "color-tone" as const,
+          })),
+        });
+      }
+      break;
+    }
+
+    case "targets-color": {
+      if (allChordMembers.length > 0) {
+        cues.push({
+          kind: "land-on",
+          label: "Land on",
+          notes: allChordMembers.map(toCueNote),
+        });
+      }
+      if (colorNotesFiltered.length > 0) {
+        cues.push({
+          kind: "color-note",
+          label: colorNotesFiltered.length === 1 ? "Color note" : "Color notes",
+          notes: colorNotesFiltered.map((n) => ({
+            internalNote: n.internalNote,
+            displayNote: n.displayNote,
+            intervalName: n.intervalName,
+            role: "color-tone" as const,
+          })),
+        });
+      }
+      break;
+    }
+
+    case "tension": {
+      if (allChordMembers.length > 0) {
+        cues.push({
+          kind: "land-on",
+          label: "Land on",
+          notes: allChordMembers.map(toCueNote),
+        });
+      }
+      const tensionMembers = allChordMembers.filter((e) => !e.inScale);
+      if (tensionMembers.length > 0) {
+        cues.push({
+          kind: "tension",
+          label: "Tension",
+          notes: tensionMembers.map((e) => ({
+            ...toCueNote(e),
+            role: "chord-tone-outside-scale" as const,
+            resolvesTo: findResolution(e.internalNote),
+          })),
+        });
+      }
+      break;
+    }
+  }
+
+  return cues;
+});
+
+// ---------------------------------------------------------------------------
+// Practice bar visibility + title
+// ---------------------------------------------------------------------------
+
+export const showChordPracticeBarAtom = atom((get) => {
+  const chordType = get(chordTypeAtom);
+  if (!chordType) return false;
+
+  const practiceLens = get(practiceLensAtom);
+
+  // Non-default lenses always show — user explicitly chose a practice focus.
+  if (practiceLens !== "targets-color") return true;
+
+  // For targets-color (default): suppress the bar for the diatonic simple case
+  // where the chord is fully in-scale, root is linked, and no color tones exist
+  // (nothing interesting to coach about).
+  const focusPreset = get(focusPresetAtom);
+  const hasOutsideChordMembers = get(hasOutsideChordMembersAtom);
+  const colorNotes = get(colorNotesAtom);
+  const chordRoot = get(chordRootAtom);
+  const rootNote = get(rootNoteAtom);
+
+  const isDiatonicSimpleCase =
+    focusPreset === "all" &&
+    !hasOutsideChordMembers &&
+    colorNotes.length === 0 &&
+    chordRoot === rootNote;
+
+  return !isDiatonicSimpleCase;
+});
+
+export const practiceBarTitleAtom = atom((get) => {
+  const chordType = get(chordTypeAtom);
+  const chordLabel = get(chordLabelAtom);
+  if (!chordType) return "";
+  return chordLabel ?? "";
+});
+
+// Badge is now minimal — always null (lens context shown via title + shape subtitle).
+export const practiceBarBadgeAtom = atom(() => null as string | null);
+
+// ---------------------------------------------------------------------------
+// Practice bar member rows (shared/outside split)
+// ---------------------------------------------------------------------------
+
+export const practiceBarSharedMembersAtom = atom((get) =>
+  get(allChordMembersAtom).filter((e) => e.inScale),
+);
+
+export const practiceBarOutsideMembersAtom = atom((get) =>
+  get(allChordMembersAtom).filter((e) => !e.inScale),
+);
